规格组删除

+++ lsj-shop-parent/lsj-shop-service/lsj-shop-service-xxx/src/main/java/com/baidu/shop/service/impl/SpecificationServiceImpl.java
@@ -62,6 +62,13 @@ public class SpecificationServiceImpl extends BaseApiService implements Specific
     @Override
     @Transactional
     public Result<JsonObject> removeSpecGroup(Integer id) {
+
+        Example example = new Example(SpecParamEntity.class);
+        example.createCriteria().andEqualTo("groupId",id);
+        List<SpecParamEntity> list = specParamMapper.selectByExample(example);
+
+        if(list.size() > 0) return this.setResultError("规格组不为空,不能删除");
+
         specGroupMapper.deleteByPrimaryKey(id);
         return this.setResultSuccess();
     }
